Add `#![allow(missing_docs)]` to generated "partial" mods
This makes sense as we might generated new items that don't have any
docs. Also, we don't forward any docs of the input structs. That we
could change, but I don't think it currently helps in any way. Maybe at
the point when Rust analyzer can fully see and understand the generated
module, then it would make sense to forward docs. But this is the
easier option now.