Change `template::format` to not be generic over `C`
This potentially improves compile times a bit. The only thing the
function needs is the `META` structure of `C`, which can be passed from
the outside. The function is still generic over the formatter, but only
three specific ones are passed. So maybe it's just compiled for those
three already? One might also think about using dynamic dispatch for
the formatter.

Though compile times are likely dominated by all the generated code,
especially the serde-derive code.