Make file format features non-default
It seems that most users only use a single file format and many forget
to disable default-features, leading to a uselessly inflated dependency
tree. This way around makes the library a tiny bit harder to get started
with, but I am willing to sacrifice this in order to save the world
some compile time.