Specify exact version requirements of dependencies
Without this, it could have been that our code depended on something
that those dependencies only added after, say `1.0`. That's because
Cargo by default uses the newest version for each dependency. I now
specified the lowest versions for which I know confique works. This is
basically no difference to before: the patch versions of `serde` and
`quote` and `proc-macro2` are determined by other dependencies, that
also depend on those. Anyway, with these version requirements everything
works.