Use `Option` instead of `std::option::Option` in emitted code (BREAKING)
This came out of #19. When deriving other things for the partial type,
`Option` often needs to appear verbatim. This is true for `clap`. A
verbatim `Option` can of course clash with another symbol in scope,
which would make the whole code not compile. But I decided it is fine
to require `Option` to refer to `std::option::Option`: in Rust, hardly
anyone shadows that symbol anyway due to its importance. So this is
incredibly unlikely to affect any real code.

I did consider emitting some code that would generate an easy to
understand error message in case someone violates this requirement, but
I couldn't come up with anything good. And the weird errors will drown
out the one good one anyway.